move UserProfileTypeWithCount logic
