+++ src/screens/Ticketing/Purchase/Overview/index.tsx
@@ -1,7 +1,11 @@
 import FullScreenHeader from '@atb/components/screen-header/full-header';
 import {useFirestoreConfiguration} from '@atb/configuration/FirestoreConfigurationContext';
 import {usePreferences} from '@atb/preferences';
-import {TariffZone} from '@atb/reference-data/types';
+import {
+  PreassignedFareProduct,
+  TariffZone,
+  UserProfile,
+} from '@atb/reference-data/types';
 import {productIsSellableInApp} from '@atb/reference-data/utils';
 import {StyleSheet} from '@atb/theme';
 import {PurchaseOverviewTexts, useTranslation} from '@atb/translations';
@@ -10,11 +14,7 @@ import {ScrollView, View} from 'react-native';
 import {TariffZoneWithMetadata} from '../TariffZones';
 import {UserProfileWithCount} from '../Travellers/use-user-count-state';
 import {TicketPurchaseScreenProps} from '../types';
-import {
-  UserProfileTypeWithCount,
-  useTariffZoneFromLocation,
-  useTravellersWithPreselectedCounts,
-} from '../utils';
+import {useTariffZoneFromLocation} from '../utils';
 import TicketDetailsSelection from './TicketDetailsSelection';
 
 type OverviewProps = TicketPurchaseScreenProps<'PurchaseOverview'>;
@@ -95,6 +95,48 @@ const PurchaseOverview: React.FC<OverviewProps> = ({
     </View>
   );
 };
+
+type UserProfileTypeWithCount = {
+  userTypeString: string;
+  count: number;
+};
+
+function getCountIfUserIsIncluded(
+  u: UserProfile,
+  selections: UserProfileTypeWithCount[],
+): number {
+  const selectedUser = selections.filter(
+    (up: UserProfileTypeWithCount) => up.userTypeString === u.userTypeString,
+  );
+
+  if (selectedUser.length < 1) return 0;
+  return selectedUser[0].count;
+}
+
+/**
+ * Get the default user profiles with count. If a default user profile has been
+ * selected in the preferences that profile will have a count of one. If no
+ * default user profile preference exists then the first user profile will have
+ * a count of one.
+ */
+export function useTravellersWithPreselectedCounts(
+  userProfiles: UserProfile[],
+  defaultSelections: UserProfileTypeWithCount[],
+  preassignedFareProduct: PreassignedFareProduct,
+) {
+  return useMemo(
+    () =>
+      userProfiles
+        .filter((u) =>
+          preassignedFareProduct.limitations.userProfileRefs.includes(u.id),
+        )
+        .map((u) => ({
+          ...u,
+          count: getCountIfUserIsIncluded(u, defaultSelections),
+        })),
+    [userProfiles, preassignedFareProduct],
+  );
+}
 
 /**
  * Get the default tariff zone, either based on current location or else the
